Remove WithPaused
WithPaused should have never existed and with the current state of poller it is
not needed any more. The poller is paused automatically.

Using WithPaused directly proves that you are doing something wrong. 0MQ is not
supposed to be used in a way that would require WithPaused to be used. If you
need to write to a socket from another thread, then connect the threads with
PAIR and send messages between them.